Make it easier to add a result to diagnostics
