+++ compiler/hash-typecheck/src/new/passes/ast_pass.rs
@@ -2,6 +2,7 @@ use hash_ast::{
     ast::{AstNodeRef, BodyBlock, Module, OwnsAstNode},
     node_map::SourceRef,
 };
+use hash_source::location::{SourceLocation, Span};
 
 use crate::new::{
     diagnostics::error::TcResult, environment::tc_env::AccessToTcEnv, ops::common::CommonOps,
@@ -20,5 +21,27 @@ pub trait AstPass: AccessToTcEnv {
             SourceRef::Module(module_source) => self.pass_module(module_source.node_ref()),
         };
         self.tc_env().try_or_add_error(result);
+    }
+
+    /// Create a [SourceLocation] from a [Span].
+    fn source_location(&self, span: Span) -> SourceLocation {
+        SourceLocation { span, id: self.current_source_info().source_id }
+    }
+
+    /// Create a [SourceLocation] at the given [hash_ast::ast::AstNode].
+    fn node_location<N>(&self, node: AstNodeRef<N>) -> SourceLocation {
+        let node_span = node.span();
+        self.source_location(node_span)
+    }
+
+    /// Unwrap a [`TcResult`] or add the error to the diagnostics.
+    fn unwrap_or_diagnostic<T>(&self, result: TcResult<T>) -> Option<T> {
+        match result {
+            Ok(ok) => Some(ok),
+            Err(err) => {
+                self.diagnostics().add_error(err);
+                None
+            }
+        }
     }
 }
